Added the ResultPlotWindow and change RunningDialog to RunningWindow.
Added the ResultPlotWindow to show the result in a chart only when the
problem is of one objectives or two.
Added the RunningWindow.fxml
Delete RunningDialog and modify is controller to load a view from fxml.
Change the name of RunningDialogController a RunningWindowController.
Modified diagrama and "documento de diseño" to show the change between
RunningDialog to RunningWindow

+++ JHawanetFramework/src/main/java/model/metaheuristic/problem/InversionCostProblem.java
@@ -185,7 +185,7 @@ public class InversionCostProblem implements Problem<IntegerSolution> {
 		Collection<Link> links = network.getLinks();
 		int i = 0;
 		for (Link link : links) {
-			double diameter = (float) this.gamas.get(iSolution.getVariable(i) - 1).getDiameter();
+			double diameter = this.gamas.get(iSolution.getVariable(i) - 1).getDiameter();
 			/*
 			 * Only pipe and valve has diameter. For this only in that elements the diameter
 			 * is set. If link is a Pump so the diameter of this is ignored.
